idb2pat infinite loop if function size < 32
`func.endEA - func.startEA` can be less than 32 (line 261), in which case `zrange` will loop infinitely since the sentinel value in `iter` will never be reached. Occurred when I used this on an ARM binary. To stop this from biting others, I added a check that will return empty list if `end < start`, which is the behavior of `range` and `xrange`.